# modules/openapi/src/commonMain/kotlin/com/sphereon/oid/fed/openapi/admin-server.yaml
openapi: 3.0.3
info:
  title: OpenID Federation API
  description: This API allows the interaction with federation endpoints to manage and verify entities within a federation.
  contact:
    name: Sphereon
    email: info@sphereon.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0-d40
servers:
  - description: SwaggerHub API Auto Mocking
    url: 'https://virtserver.swaggerhub.com/SphereonInt/OpenIDFederationAPI/1.0.0-d40'
  - url: http://localhost:8081
    description: Admin Server
paths:
...
  '/trust-mark-types':
    get:
      summary: Get all Trust Mark Types
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      responses:
        '200':
          description: List of trust mark types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrustMarkType'

    post:
      summary: Create a Trust Mark Type
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrustMarkType'
      responses:
        '201':
          description: Trust mark definition created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustMarkType'
  '/trust-mark-types/{id}':
    get:
      summary: Get a Trust Mark Type by ID
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Trust mark definition details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustMarkType'
    delete:
      summary: Delete a Trust Mark Type
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Trust mark definition deleted
  /trust-mark-types/{id}/issuers:
    get:
      summary: Get Issuers for a Trust Mark Type
      description: Retrieve a list of issuers for the specified Trust Mark Type.
      operationId: getIssuersForTrustMarkType
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Identifier of the Trust Mark Type.
      responses:
        '200':
          description: A list of issuer identifiers.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: https://issuer.example.com
        '404':
          description: Account or Trust Mark Type not found.
    post:
      summary: Add Issuer to Trust Mark Type
      description: Add a new issuer to the specified Trust Mark Type.
      operationId: addIssuerToTrustMarkType
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Identifier of the Trust Mark Type.
      requestBody:
        description: Details of the issuer to be added.
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - identifier
              properties:
                identifier:
                  type: string
                  description: Identifier of the issuer.
                  example: https://issuer.example.com
      responses:
        '201':
          description: Issuer successfully added to the Trust Mark Type.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: Unique ID of the Trust Mark Issuer.
                    example: issuer-123
                  identifier:
                    type: string
                    description: Identifier of the issuer.
                    example: https://issuer.example.com
        '404':
          description: Account or Trust Mark Type not found.
        '400':
          description: Invalid request body.
  /trust-mark-types/{id}/issuers/{issuerIdentifier}:
    delete:
      summary: Remove Issuer from Trust Mark Type
      description: Remove an issuer from the specified Trust Mark Type.
      operationId: removeIssuerFromTrustMarkType
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Identifier of the Trust Mark Type.
        - name: issuerIdentifier
          in: path
          required: true
          schema:
            type: string
          description: Identifier of the issuer to be removed.
      responses:
        '200':
          description: Issuer successfully removed from the Trust Mark Type.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: Unique ID of the removed Trust Mark Issuer.
                    example: issuer-123
                  identifier:
                    type: string
                    description: Identifier of the removed issuer.
                    example: https://issuer.example.com
        '404':
          description: Account, Trust Mark Type, or Issuer not found.
  '/trust-marks':
    post:
      summary: Create or Update a Trust Mark
      description: Create or update a Trust Mark for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                dry_run: # TO-DO Add correct required attributes
                  type: boolean
                  description: 'If true, the entity statement will be generated but not persisted.'
                  default: false
      responses:
        '200':
          description: Trust Mark dry-run successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  trustMarkId:
                    type: string
                    description: The identifier of the created or updated Trust Mark.
        '201':
          description: Trust Mark created or updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  trustMarkId:
                    type: string
                    description: The identifier of the created or updated Trust Mark.
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List Trust Marks
      description: List all Trust Marks for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
      responses:
        '200':
          description: Successful fetch of Trust Marks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    trustMarkId:
                      type: string
                      description: The identifier of the Trust Mark.
                    trustMark:
                      type: string
                      description: The JWT of the Trust Mark.
                    entityId:
                      type: string
                      description: The Entity Identifier of the entity to which the Trust Mark is issued.
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  '/trust-marks/{trustMarkId}':
    delete:
      summary: Delete a Trust Mark
      description: Delete an existing Trust Mark for the specified account.
      parameters:
        - $ref: '#/components/parameters/AccountUsername'
        - name: trustMarkId
          in: path
          required: true
          schema:
            type: string
          description: The identifier of the Trust Mark to be deleted.
      responses:
        '200':
          description: Trust Mark deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Trust Mark deleted successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Trust Mark not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    AccountUsername:
      name: X-Account-Username
      in: header
      description: Optional account username to operate on behalf of
      required: false
      schema:
        type: string
...
  schemas:
...
    CreateTrustMarkType:
      type: object
      x-tags:
        - federation
      properties:
        identifier:
          type: string
          description: The unique identifier for the Trust Mark Type.
          example: "https://www.example.com/oidf/trustmark/underageSafetyVerified"
      required:
        - identifier
...
    ErrorResponse:
      type: object
      x-tags:
        - federation
      required:
        - error
        - message
      properties:
        status:
          type: integer
          x-stoplight:
            id: joax1cgth4uzd
        error:
          type: string
        message:
          type: string
          description: A human-readable short text describing the error.
        timestamp:
          type: integer
          x-stoplight:
            id: qtn6mqbzpjctv
...
    TrustMarkType:
      type: object
      x-tags:
        - federation
      properties:
        id:
          type: integer
          description: The unique identifier of the Trust Mark Type.
          example: 123
        identifier:
          type: string
          description: The unique identifier for the Trust Mark Type.
          example: "https://www.example.com/oidf/trustmark/underageSafetyVerified"
        createdAt:
          type: string
          format: date-time
          description: The timestamp when the Trust Mark Type was created.
          example: "2024-12-01T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: The timestamp when the Trust Mark Type was last updated.
          example: "2024-12-15T15:30:00Z"
          nullable: true
      required:
        - id
        - identifier
        - name
        - issuerPolicy
        - createdAt